CONTROL PRESUPUESTO: Variacion de insumos 01022018

+++ resources/views/control_presupuesto/cambio_presupuesto/variacion_volumen.blade.php
@@ -87,7 +87,7 @@
                                     <td>@{{ concepto.sector }}</td>
                                     <td>@{{ concepto.cuadrante }}</td>
                                     <td>@{{ concepto.unidad }}</td>
-                                    <td class="text-right">@{{ parseInt(concepto.cantidad_presupuestada).formatMoney(2, ',','.') }}</td>
+                                    <td class="text-right">@{{ parseFloat(concepto.cantidad_presupuestada).formatMoney(2, ',','.') }}</td>
                                     <td>
                                         <div class="form-group" :class="{'has-error': validation_errors.has('form_save_solicitud.Cantidad Presupuestada Nueva ' + (i+1))}">
                                             <input type="text" :name="'Cantidad Presupuestada Nueva ' + (i+1)" v-validate="'decimal|regex:[^0]+'" class="form-control input-sm" v-model="concepto.cantidad_presupuestada_nueva">
